Moving and renaming AVC and tests, adding more docstring examples
