+++ tests/unit/core/test_annual_value_calculator.py
@@ -9,224 +9,6 @@ from numpy.typing import NDArray
 
 
 @pytest.mark.parametrize(
-    argnames="datetimes, growing_season,  as_acclim, context_manager, error_message",
-    argvalues=(
-        pytest.param(
-            np.arange(
-                np.datetime64("2000-01-01"),
-                np.datetime64("2010-01-01"),
-                np.timedelta64(30, "m"),
-            ),
-            np.ones(
-                (365 * 7 + 366 * 3) * 48, dtype=np.bool_
-            ),  # 10 years of 30 min obs, with 3 leap years
-            True,
-            does_not_raise(),
-            None,
-            id="acclim_model_good",
-        ),
-        pytest.param(
-            np.arange(
-                np.datetime64("2000-01-01"),
-                np.datetime64("2010-01-01"),
-                np.timedelta64(30, "m"),
-            ),
-            np.ones(
-                (365 * 7 + 366 * 3) * 48, dtype=np.bool_
-            ),  # 10 years of 30 min obs, with 3 leap years
-            False,
-            does_not_raise(),
-            None,
-            id="datetimes_half_hourly_good",
-        ),
-        pytest.param(
-            np.arange(
-                np.datetime64("2000-01-01"),
-                np.datetime64("2010-01-01"),
-                np.timedelta64(30, "m"),
-            ).astype(np.int_),
-            np.ones(
-                (365 * 5 + 366 * 3) * 48, dtype=np.bool_
-            ),  # 10 years of 30 min obs, with 3 leap years
-            False,
-            pytest.raises(ValueError),
-            "The timings argument must be an AcclimationModel "
-            "or an array of datetime64 values",
-            id="timings_not_acclim_or_datetimes",
-        ),
-        pytest.param(
-            np.arange(
-                np.datetime64("2000-01-01"),
-                np.datetime64("2010-01-01"),
-                np.timedelta64(30, "m"),
-            ),
-            np.ones(
-                (365 * 5 + 366 * 3) * 48, dtype=np.bool_
-            ),  # 10 years of 30 min obs, with 3 leap years
-            False,
-            pytest.raises(ValueError),
-            "Growing season data is not the same shape as the timing data",
-            id="bad_growing_season_shape",
-        ),
-        pytest.param(
-            np.arange(
-                np.datetime64("2000-01-01"),
-                np.datetime64("2010-01-01"),
-                np.timedelta64(30, "m"),
-            ),
-            np.ones(
-                (365 * 7 + 366 * 3) * 48,
-            ),  # 10 years of 30 min obs, with 3 leap years
-            False,
-            pytest.raises(ValueError),
-            "Growing season data is not an array of boolean values",
-            id="bad_growing_season_dtype",
-        ),
-        pytest.param(
-            np.arange(
-                np.datetime64("2000-01-31"),
-                np.datetime64("2009-11-30"),
-                np.timedelta64(30, "m"),
-            ),
-            np.ones(
-                (365 * 7 + 366 * 3) * 48, dtype=np.bool_
-            ),  # 10 years of 30 min obs, with 3 leap years
-            True,
-            pytest.raises(ValueError),
-            "Data timings do not cover complete years to within tolerance",
-            id="acclim_model_incomplete_years",
-        ),
-        pytest.param(
-            np.arange(
-                np.datetime64("2000-01-01"),
-                np.datetime64("2010-01-01"),
-                np.timedelta64(1, "D"),
-            ),
-            np.ones(
-                (365 * 7 + 366 * 3), dtype=np.bool_
-            ),  # 10 years of daily obs, with 3 leap years
-            False,
-            does_not_raise(),
-            None,
-            id="datetimes_daily_good",
-        ),
-        pytest.param(
-            np.arange(
-                np.datetime64("2000-01-01"),
-                np.datetime64("2010-01-01"),
-                np.timedelta64(1, "W"),
-            ),
-            np.ones(522, dtype=np.bool_),  # 10 years of weekly obs
-            False,
-            does_not_raise(),
-            None,
-            id="datetimes_weekly_good",
-        ),
-        pytest.param(
-            np.arange(
-                np.datetime64("2000-01-05"),
-                np.datetime64("2010-01-01"),
-                np.timedelta64(1, "W"),
-            ),
-            np.ones(522, dtype=np.bool_),  # 10 years of weekly obs
-            False,
-            does_not_raise(),
-            None,
-            id="datetimes_weekly_short_but_inside_tolerance_good",
-        ),
-        pytest.param(
-            np.arange(
-                np.datetime64("2000-01-08"),
-                np.datetime64("2010-01-01"),
-                np.timedelta64(1, "W"),
-            ),
-            np.ones(522, dtype=np.bool_),  # 10 years of weekly obs
-            False,
-            does_not_raise(),
-            None,
-            id="datetimes_weekly_short_outside_tolerance",
-        ),
-    ),
-)
-def test_AnnualValueCalculator_init(
-    datetimes, as_acclim, growing_season, context_manager, error_message
-):
-    """Test failure modes and success modes for initialising AVC instances."""
-    from pyrealm.phenology.fapar_limitation import AnnualValueCalculator
-    from pyrealm.pmodel.acclimation import AcclimationModel
-
-    if as_acclim:
-        datetimes = AcclimationModel(datetimes=datetimes)
-
-    with context_manager as cmgr:
-        _ = AnnualValueCalculator(timing=datetimes, growing_season=growing_season)
-        return
-
-    assert str(cmgr.value) == error_message
-
-
-@pytest.mark.parametrize(
-    argnames="function, within_growing_season, expected",
-    argvalues=(
-        pytest.param(
-            "mean",
-            True,
-            np.repeat(1, 10),
-            id="mean",
-        ),
-        pytest.param(
-            "sum",
-            True,
-            np.repeat(200, 10),
-            id="sum",
-        ),
-        pytest.param(
-            "sum",
-            False,
-            np.array([366, 365, 365, 365, 366, 365, 365, 365, 366, 365]),
-            id="sum",
-        ),
-    ),
-)
-def test_AnnualValueCalculator(function, within_growing_season, expected):
-    """Basic tests of the AnnualValueCalculator.
-
-    This checks that the annual value calculator extracts the 1 values successfully from
-    within the daily data and then correctly summarises them for different functions and
-    with and without the growing season.
-    """
-
-    from pyrealm.phenology.fapar_limitation import AnnualValueCalculator
-
-    # Set up a 10 year period
-    datetimes = np.arange(
-        np.datetime64("2000-01-01"),
-        np.datetime64("2010-01-01"),
-        np.timedelta64(30, "m"),
-    )
-
-    # Set up a growing season that includes 200 days in the middle of each year
-    _, obs_in_year = np.unique(datetimes.astype("datetime64[Y]"), return_counts=True)
-    days_in_year = obs_in_year / 48
-    growing_season = np.concatenate(
-        [np.repeat([False, True, False], (75, 200, n - 275)) for n in days_in_year]
-    )
-
-    # Values - has ones at midday for all days
-    values = np.tile(np.repeat((0, 1, 0), (24, 1, 23)), 3653)
-
-    # Create the calculator instance
-    avc = AnnualValueCalculator(timing=datetimes, growing_season=growing_season)
-
-    # Apply the method and check the outcome.
-    result = avc.get_annual_values(
-        values=values, function=function, within_growing_season=within_growing_season
-    )
-
-    assert_allclose(result, expected)
-
-
-@pytest.mark.parametrize(
     argnames="datetimes, growing_season,  as_acclim, endpoint,"
     " context_manager, error_message, year_completeness",
     argvalues=(
@@ -445,7 +227,7 @@ def test_AnnualValueCalculator(function, within_growing_season, expected):
         ),
     ),
 )
-def test_AnnualValueCalculatorMarkII_init(
+def test_AnnualValueCalculator_init(
     datetimes,
     as_acclim,
     growing_season,
@@ -455,14 +237,14 @@ def test_AnnualValueCalculatorMarkII_init(
     year_completeness,
 ):
     """Test failure modes and success modes for initialising AVC instances."""
-    from pyrealm.phenology.fapar_limitation import AnnualValueCalculatorMarkII
+    from pyrealm.core.time_series import AnnualValueCalculator
     from pyrealm.pmodel.acclimation import AcclimationModel
 
     if as_acclim:
         datetimes = AcclimationModel(datetimes=datetimes)
 
     with context_manager as cmgr:
-        avc = AnnualValueCalculatorMarkII(
+        avc = AnnualValueCalculator(
             timing=datetimes, growing_season=growing_season, endpoint=endpoint
         )
 
@@ -596,7 +378,7 @@ GROWING_SEASON = np.tile(np.repeat([0, 1, 0], [3, 6, 3]), 10).astype(np.bool_)
         ),
     ),
 )
-def test_AnnualValueCalculatorMarkII_get_annual(
+def test_AnnualValueCalculator_get_annual(
     init,
     values,
     total_expected,
@@ -605,9 +387,9 @@ def test_AnnualValueCalculatorMarkII_get_annual(
     mean_expected_within_gs,
 ):
     """Test the calculation of annual totals and means."""
-    from pyrealm.phenology.fapar_limitation import AnnualValueCalculatorMarkII
+    from pyrealm.core.time_series import AnnualValueCalculator
 
-    avc = AnnualValueCalculatorMarkII(**init)
+    avc = AnnualValueCalculator(**init)
 
     calculated = avc.get_annual_totals(values, within_growing_season=False)
     assert_allclose(calculated, total_expected)
